Fix arithmetic UB in decorator tasks
Three arithmetic undefined behavior vulnerabilities fixed:

1. BTRepeat: set_times() now clamps to [1, 1000000] to prevent signed
   integer overflow in cur_iteration. Without this, values approaching
   INT_MAX cause UB when cur_iteration is incremented past INT_MAX
   during tick execution.

2. BTRunLimit: Changed run_limit and num_runs from int to int64_t.
   An int counter overflows after ~410 days at 60 FPS on dedicated
   servers. int64_t extends this to ~9.7 trillion years. Also enforces
   minimum value of 1 via MAX(1, p_value) in the setter.

3. BTForEach: Changed current_idx from int to int64_t to match
   Array::size() return type in GDExtension (which returns int64_t).
   The int/int64_t type mismatch caused implicit narrowing conversion
   UB when array sizes exceed INT_MAX in GDExtension builds.

+++ bt/tasks/decorators/bt_for_each.h
@@ -22,7 +22,7 @@ private:
 	StringName array_var;
 	StringName save_var;
 
-	int current_idx;
+	int64_t current_idx; // Match Array::size() return type (int64_t in GDExtension)
 
 protected:
 	static void _bind_methods();
